Add rate limit for downloads
Some websites limit the number of accesses to it and it should also be
in the interesst of all to be good netizens and not overly stress pages.

Allow the user to restict the access to pages to x per minute, whereas
values smaller then 1 result in less then one access per minute.
0.5 will result in one access every two minutes, reducing strain on the
webservers we get our novels from.

Default has been set to 1 per second, which is quite high, but in the
spirit of the previous implementation which did not throttle at all.

Signed-off-by: Fabian Pflug <fabian.pflug@gumulka.eu>