Allow for fast breakout of ratelimit with CTRL+C
If the ratelimit was higher then 1/second, the process would sleep for
the time defined in ratelimit. Which could be 1 minute on paranoid
sites. Break the sleep into small chucks of half a second. Which was
choosen as a good middle ground betwen constantly iterating through the
loop and not be reactive at all.

Signed-off-by: Fabian Pflug <fabian.pflug@gumulka.eu>